Fix of a sleeping bug, other bugfixes, QoL

+++ inMyRoom_vulkan/src/PrimitivesMaterials.cpp
@@ -55,9 +55,9 @@ PrimitivesMaterials::PrimitivesMaterials(tinygltf::Model& in_model, MaterialsTex
                                                                                        dsg_create_infos_ptr,
                                                                                        false); /* in_releaseable_sets */
 
-        for (size_t set_index = 0; set_index < textures_bindings.size(); set_index++)
+        for (uint32_t set_index = 0; set_index < textures_bindings.size(); set_index++)
         {
-            for (size_t binding_index = 0; binding_index < textures_bindings[set_index].size(); binding_index++)
+            for (uint32_t binding_index = 0; binding_index < textures_bindings[set_index].size(); binding_index++)
             {
                 texturesOfMaterialsDescriptorSetGroup_uptr->set_binding_item(set_index,
                                                                              binding_index,
